requirements.txt
fastmcp>=0.1.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
pydantic>=2.0.0
